+++ src/main/java/com/example/spring_homework3/dao/AccountJpaRepository.java
@@ -6,10 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-public interface AccountJpaRepository extends JpaRepository<Account, Integer> {
+public interface AccountJpaRepository extends JpaRepository<Account, Long> {
     Account findByNumber(String number);
     @Modifying
     @Transactional
